gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
asyncpg==0.29.0
qdrant-client==1.11.0
minio==7.2.8
google-generativeai==0.7.2
pdfplumber==0.11.0
pdf2image==1.17.0
python-multipart==0.0.9
pyyaml==6.0.1
python-dotenv==1.0.1
